% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_utils.R
\name{nonull}
\alias{nonull}
\title{Exclude NULL elements from a list}
\usage{
nonull(x)
}
\arguments{
\item{x}{A list whose NULL elements will be filtered}
}
\description{
Exclude NULL elements from a list
}
\note{
adapted by
\url{https://github.com/rstudio/leaflet/blob/master/R/utils.R}
}
