% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papayaWidget.R
\name{papaya}
\alias{papaya}
\title{Creating a Papaya widget}
\usage{
papaya(img = NULL, elementId = NULL, width = NULL, height = NULL,
  options = NULL, sync_view = FALSE, hide_toolbar = FALSE,
  hide_controls = FALSE, orthogonal = TRUE)
}
\arguments{
\item{img}{Vector of character file names or list of \code{nifti} images}

\item{elementId}{ID for the element in the DOM}

\item{width}{Width of the widget}

\item{height}{height of the widget}

\item{options}{to pass to papaya widget.  Should be a list with
the same number of elements as images passed in.  Options can
be generated using \code{\link{papayaOptions}}}

\item{sync_view}{should the view be synced to other Papaya viewers?}

\item{hide_toolbar}{Hide the toolbar for this viewer?}

\item{hide_controls}{Hide the controls (increment) for this viewer?}
}
\description{
Wraps a widget for the Papaya JavaScript library
}
\examples{
img = kirby21.t1::get_t1_filenames()[1]

papaya(img, options = papayaOptions(min = 0, max = 5e5))
papaya(c(img, img),
options = list(papayaOptions(alpha = 0.5),
papayaOptions(alpha = 0.5, lut = "Red Overlay"))
)


}
