% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papayaOptions.R
\name{papayaOptions}
\alias{papayaOptions}
\alias{validateOptions}
\title{Options for Papaya Image}
\usage{
papayaOptions(
  alpha = NULL,
  icon = NULL,
  interpolation = NULL,
  lut = NULL,
  negative_lut = NULL,
  max = NULL,
  maxPercent = NULL,
  min = NULL,
  minPercent = NULL,
  rotation = NULL,
  rotationPoint = NULL,
  parametric = NULL,
  symmetric = NULL
)

validateOptions(...)
}
\arguments{
\item{alpha}{The overlay image alpha level (0 to 1).}

\item{icon}{Image url to use as the toolbar icon (rescaled to 18x18).}

\item{interpolation}{If true, image will have a smooth display
(default true).}

\item{lut}{The color table name (built-in or custom).}

\item{negative_lut}{The color table name (built-in or custom)
used by the negative side of the parametric pair.}

\item{max}{The display range maximum.}

\item{maxPercent}{The display range maximum as a percentage of
image max.}

\item{min}{The display range minimum.}

\item{minPercent}{The display range minimum as a percentage of
image max.}

\item{rotation}{An array of rotation degrees about each of the
three axes (e.g., [0,0,45] = 45 degree rotation about Z axis)}

\item{rotationPoint}{The point of rotation. Options are "center"
(default), "origin", or "crosshairs".}

\item{parametric}{If true, loads two views of the same image,
one in the positive and one in the negative range. (Also see
combineParametric.)}

\item{symmetric}{If true, sets the negative range of a parametric
pair to the same size as the positive range.}

\item{...}{additional arguments (not used)}
}
\description{
Options for Papaya Image
}
\examples{
papayaOptions(alpha = 0.5)
papayaOptions()
papayaOptions(rotation = c(0, 0, 45),
parametric = TRUE, rotationPoint = "center")
}
