% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papayaWidget.R
\name{papaya}
\alias{papaya}
\title{Creating a Papaya widget}
\usage{
papaya(
  img = NULL,
  elementId = NULL,
  width = NULL,
  height = NULL,
  options = NULL,
  sync_view = FALSE,
  hide_toolbar = FALSE,
  hide_controls = FALSE,
  orthogonal = TRUE,
  interpolation = TRUE,
  title = NULL
)
}
\arguments{
\item{img}{Vector of character file names or list of \code{nifti} images}

\item{elementId}{ID for the element in the DOM}

\item{width}{Width of the widget}

\item{height}{height of the widget}

\item{options}{to pass to papaya widget.  Should be a list with
the same number of elements as images passed in.  Options can
be generated using \code{\link{papayaOptions}}}

\item{sync_view}{should the view be synced to other Papaya viewers?}

\item{hide_toolbar}{Hide the toolbar for this viewer?}

\item{hide_controls}{Hide the controls (increment) for this viewer?}

\item{orthogonal}{Should an orthogonal view be displayed (\code{TRUE})
versus just a slice (\code{FALSE}).}

\item{interpolation}{Should smoothing/interpolation be done in the viewer?
This can be overridden if individual image options set \code{interpolation}
as well}

\item{title}{the title for the overlay itself}
}
\description{
Wraps a widget for the Papaya JavaScript library
}
\examples{
papaya(interpolation = TRUE)
papaya(interpolation = FALSE)

img = kirby21.t1::get_t1_filenames()[1]

papaya(img, options = papayaOptions(min = 0, max = 8e5))
papaya(c(img, img),
options = list(papayaOptions(alpha = 0.5),
papayaOptions(alpha = 0.5, lut = "Red Overlay"))
)

\dontrun{
papaya(img, options = papayaOptions(min = 0, max = 8e5))
papaya(c(img, img),
options = list(papayaOptions(alpha = 0.5),
papayaOptions(alpha = 0.5, lut = "Red Overlay"))
)
nii = neurobase::readnii(img)
nii = nii < 3e5 & nii > 1e5
papaya(list(img, nii), options = list(list(), papayaOptions(interpolation = FALSE)))
}

}
