% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papayaWidget.R
\name{papayaWidget}
\alias{papayaWidget}
\title{<Add Title>}
\usage{
papayaWidget(img, passingMethod = c("embed", "file"), elementId = NULL,
  width = NULL, height = NULL)
}
\description{
<Add Description>
}
